feat: v2 president and legislator pipeline for town and village
add a new feature to generate president data for v2, and finish the implementation of pipeline to generate town and village data for legislator.

something should be careful is that the candidates data in v2 is got from GQL query to WHORU_GQL_PROD(because no data in dev) and it's async.

TODO: (1) Merge some implementation in pipeline (2) Other V2 data